+++ src/lib/Scenes/Artwork/Artwork.tsx
@@ -5,7 +5,7 @@ import Separator from "lib/Components/Separator"
 import { defaultEnvironment } from "lib/relay/createEnvironment"
 import renderWithLoadProgress from "lib/utils/renderWithLoadProgress"
 import React from "react"
-import { Dimensions, FlatList } from "react-native"
+import { FlatList } from "react-native"
 import { createFragmentContainer, graphql, QueryRenderer } from "react-relay"
 import { AboutArtistFragmentContainer as AboutArtist } from "./Components/AboutArtist"
 import { AboutWorkFragmentContainer as AboutWork } from "./Components/AboutWork"
@@ -86,7 +86,7 @@ export class Artwork extends React.Component<Props> {
   render() {
     return (
       <Theme>
-        <Box pt={4} pb={2}>
+        <Box pb={2}>
           <FlatList
             data={this.sections()}
             ItemSeparatorComponent={() => (
@@ -95,7 +95,9 @@ export class Artwork extends React.Component<Props> {
               </Box>
             )}
             keyExtractor={(item, index) => item.type + String(index)}
-            renderItem={item => <Box px={2}>{this.renderItem(item)}</Box>}
+            renderItem={item =>
+              item.item === "header" ? this.renderItem(item) : <Box px={2}>{this.renderItem(item)}</Box>
+            }
           />
         </Box>
       </Theme>
@@ -163,7 +165,7 @@ export const ArtworkRenderer: React.SFC<{ artworkID: string }> = ({ artworkID })
     <QueryRenderer<ArtworkQuery>
       environment={defaultEnvironment}
       query={graphql`
-        query ArtworkQuery($artworkID: String!, $excludeArtworkIds: [String!], $screenWidth: Int!) {
+        query ArtworkQuery($artworkID: String!, $excludeArtworkIds: [String!]) {
           artwork(id: $artworkID) {
             ...Artwork_artwork
           }
@@ -171,7 +173,6 @@ export const ArtworkRenderer: React.SFC<{ artworkID: string }> = ({ artworkID })
       `}
       variables={{
         artworkID,
-        screenWidth: Dimensions.get("screen").width,
         excludeArtworkIds: [artworkID],
       }}
       render={renderWithLoadProgress(ArtworkContainer)}